created useGraphQuery hook
